requirements.txt
torchmonarch
transformers
accelerate
datasets
huggingface_hub
scipy
deepeval
sentencepiece
tiktoken
matplotlib